requirements.txt
redis
psycopg2-binary

